@@ -1,6 +1,5 @@
 package org.teamapps.demolessons.p1_intro.l01_panel;
 
-import com.google.common.io.Files;
 import org.teamapps.common.format.Color;
 import org.teamapps.demolessons.DemoLesson;
 import org.teamapps.icon.material.MaterialIcon;
@@ -8,11 +7,13 @@ import org.teamapps.server.jetty.embedded.TeamAppsJettyEmbeddedServer;
 import org.teamapps.ux.component.Component;
 import org.teamapps.ux.component.dummy.DummyComponent;
 import org.teamapps.ux.component.panel.Panel;
+import org.teamapps.ux.component.rootpanel.RootPanel;
 import org.teamapps.ux.session.SessionContext;
-import org.teamapps.webcontroller.SimpleWebController;
+import org.teamapps.webcontroller.WebController;
 
 public class PanelDemo implements DemoLesson {
 
+    // Field rootComponent is the Content of this Demo
     private Component rootComponent;
 
     // Constructor
@@ -29,16 +30,25 @@ public class PanelDemo implements DemoLesson {
         panel.setContent(new DummyComponent());
     }
 
+    // DemoLesson Interface method
+    // returns the content of the Demo
     public Component getRootComponent(){
         return rootComponent;
     }
 
+    // main method to launch the Demo standalone
     public static void main(String[] args) throws Exception {
-        SimpleWebController controller = new SimpleWebController(context -> {
+        WebController controller = sessionContext -> {
 
-            // SimpleWebController requires a Component
-            return new PanelDemo(context).getRootComponent();
-        });
-        new TeamAppsJettyEmbeddedServer(controller, Files.createTempDir()).start();
+            // RootPanel is the global container of every element
+            RootPanel rootPanel = new RootPanel();
+            sessionContext.addRootPanel(null, rootPanel);
+
+            // get the rootComponent of the Demo and set it as the content of the global container (rootPanel)
+            Component rootComponent = new PanelDemo(sessionContext).getRootComponent();
+            rootPanel.setContent(rootComponent);
+
+        };
+        new TeamAppsJettyEmbeddedServer(controller).start();
     }
 }
