@@ -1,20 +1,19 @@
 package org.teamapps.demolessons.p1_intro.l04_richtexteditor;
 
-import com.google.common.io.Files;
 import org.teamapps.demolessons.DemoLesson;
 import org.teamapps.icon.material.MaterialIcon;
 import org.teamapps.server.jetty.embedded.TeamAppsJettyEmbeddedServer;
 import org.teamapps.ux.component.Component;
-import org.teamapps.ux.component.dummy.DummyComponent;
 import org.teamapps.ux.component.field.richtext.RichTextEditor;
 import org.teamapps.ux.component.flexcontainer.VerticalLayout;
 import org.teamapps.ux.component.panel.Panel;
+import org.teamapps.ux.component.rootpanel.RootPanel;
 import org.teamapps.ux.session.SessionContext;
-import org.teamapps.webcontroller.SimpleWebController;
+import org.teamapps.webcontroller.WebController;
 
 public class RichTextEditorDemo implements DemoLesson {
 
-    private Component rootComponent = new DummyComponent();
+    private Component rootComponent;
     private SessionContext context;
 
     public RichTextEditorDemo(SessionContext context) {
@@ -27,9 +26,7 @@ public class RichTextEditorDemo implements DemoLesson {
 
         RichTextEditor textField2 = new RichTextEditor();
 
-        textField1.onValueChanged.addListener(text -> {
-            textField2.setValue("New Content: " + text);
-        });
+        textField1.onValueChanged.addListener(text -> textField2.setValue("New Content: " + text));
 
         VerticalLayout verticalLayout = new VerticalLayout();
         verticalLayout.addComponent(textField1);
@@ -46,14 +43,20 @@ public class RichTextEditorDemo implements DemoLesson {
     public void handleDemoSelected() { }
 
 
+    // main method to launch the Demo standalone
     public static void main(String[] args) throws Exception {
+        WebController controller = sessionContext -> {
+            RootPanel rootPanel = new RootPanel();
+            sessionContext.addRootPanel(null, rootPanel);
 
-        SimpleWebController controller = new SimpleWebController(context -> {
+            // create new instance of the Demo Class
+            DemoLesson demo = new RichTextEditorDemo(sessionContext);
 
-            RichTextEditorDemo textFieldDemo = new RichTextEditorDemo(context);
-            textFieldDemo.handleDemoSelected();
-            return textFieldDemo.getRootComponent();
-        });
-        new TeamAppsJettyEmbeddedServer(controller, Files.createTempDir()).start();
+            // call the method defined in the DemoLesson Interface
+            demo.handleDemoSelected();
+
+            rootPanel.setContent(demo.getRootComponent());
+        };
+        new TeamAppsJettyEmbeddedServer(controller).start();
     }
 }
