@@ -1,21 +1,20 @@
 package org.teamapps.demolessons.p1_intro.l05_label;
 
-import com.google.common.io.Files;
 import org.teamapps.demolessons.DemoLesson;
 import org.teamapps.icon.material.MaterialIcon;
 import org.teamapps.server.jetty.embedded.TeamAppsJettyEmbeddedServer;
 import org.teamapps.ux.component.Component;
-import org.teamapps.ux.component.dummy.DummyComponent;
 import org.teamapps.ux.component.field.Label;
 import org.teamapps.ux.component.field.TextField;
 import org.teamapps.ux.component.flexcontainer.VerticalLayout;
 import org.teamapps.ux.component.panel.Panel;
+import org.teamapps.ux.component.rootpanel.RootPanel;
 import org.teamapps.ux.session.SessionContext;
-import org.teamapps.webcontroller.SimpleWebController;
+import org.teamapps.webcontroller.WebController;
 
 public class LabelDemo implements DemoLesson {
 
-    private Component rootComponent = new DummyComponent();
+    private Component rootComponent;
     private SessionContext context;
 
     public LabelDemo(SessionContext context) {
@@ -32,9 +31,7 @@ public class LabelDemo implements DemoLesson {
 
         TextField textField2 = new TextField();
 
-        textField1.onTextInput.addListener(text -> {
-            textField2.setValue("Your name is: " + text);
-        });
+        textField1.onTextInput.addListener(text -> textField2.setValue("Your name is: " + text));
 
         VerticalLayout verticalLayout = new VerticalLayout();
         verticalLayout.addComponent(label1);
@@ -53,14 +50,20 @@ public class LabelDemo implements DemoLesson {
     public void handleDemoSelected() { }
 
 
+    // main method to launch the Demo standalone
     public static void main(String[] args) throws Exception {
+        WebController controller = sessionContext -> {
+            RootPanel rootPanel = new RootPanel();
+            sessionContext.addRootPanel(null, rootPanel);
 
-        SimpleWebController controller = new SimpleWebController(context -> {
+            // create new instance of the Demo Class
+            DemoLesson demo = new LabelDemo(sessionContext);
 
-            LabelDemo textFieldDemo = new LabelDemo(context);
-            textFieldDemo.handleDemoSelected();
-            return textFieldDemo.getRootComponent();
-        });
-        new TeamAppsJettyEmbeddedServer(controller, Files.createTempDir()).start();
+            // call the method defined in the DemoLesson Interface
+            demo.handleDemoSelected();
+
+            rootPanel.setContent(demo.getRootComponent());
+        };
+        new TeamAppsJettyEmbeddedServer(controller).start();
     }
 }
