@@ -1,6 +1,5 @@
 package org.teamapps.demolessons.p1_intro.l06_button;
 
-import com.google.common.io.Files;
 import org.teamapps.demolessons.DemoLesson;
 import org.teamapps.icon.material.MaterialIcon;
 import org.teamapps.server.jetty.embedded.TeamAppsJettyEmbeddedServer;
@@ -8,25 +7,22 @@ import org.teamapps.ux.component.Component;
 import org.teamapps.ux.component.dummy.DummyComponent;
 import org.teamapps.ux.component.field.Button;
 import org.teamapps.ux.component.panel.Panel;
+import org.teamapps.ux.component.rootpanel.RootPanel;
 import org.teamapps.ux.component.template.BaseTemplateRecord;
 import org.teamapps.ux.session.SessionContext;
-import org.teamapps.webcontroller.SimpleWebController;
+import org.teamapps.webcontroller.WebController;
 
 public class ButtonDemo implements DemoLesson {
 
-    private Component rootComponent = new DummyComponent();
-    private SessionContext context;
+    private Component rootComponent;
 
-    public ButtonDemo(SessionContext context) {
-        this.context = context;
+    public ButtonDemo(SessionContext sessionContext) {
 
         Panel panel = new Panel(MaterialIcon.LIGHTBULB_OUTLINE, "Button Demo");
         rootComponent = panel;
 
-        Button<BaseTemplateRecord> button = Button.create("Do it!!!     Expand it", new DummyComponent("foo"));
-        button.onValueChanged.addListener(aBoolean -> {
-            context.showNotification(MaterialIcon.FLAG, "Button clicked!");
-        });
+        Button<BaseTemplateRecord> button = Button.create("Do it! & Expand dropdownComponent of Button", new DummyComponent("foo"));
+        button.onClicked.addListener(() -> sessionContext.showNotification(MaterialIcon.FLAG, "Button clicked!"));
         panel.setContent(button);
 
     }
@@ -39,14 +35,20 @@ public class ButtonDemo implements DemoLesson {
     public void handleDemoSelected() { }
 
 
+    // main method to launch the Demo standalone
     public static void main(String[] args) throws Exception {
+        WebController controller = sessionContext -> {
+            RootPanel rootPanel = new RootPanel();
+            sessionContext.addRootPanel(null, rootPanel);
 
-        SimpleWebController controller = new SimpleWebController(context -> {
+            // create new instance of the Demo Class
+            DemoLesson demo = new ButtonDemo(sessionContext);
 
-            ButtonDemo textFieldDemo = new ButtonDemo(context);
-            textFieldDemo.handleDemoSelected();
-            return textFieldDemo.getRootComponent();
-        });
-        new TeamAppsJettyEmbeddedServer(controller, Files.createTempDir()).start();
+            // call the method defined in the DemoLesson Interface
+            demo.handleDemoSelected();
+
+            rootPanel.setContent(demo.getRootComponent());
+        };
+        new TeamAppsJettyEmbeddedServer(controller).start();
     }
 }
