@@ -1,6 +1,5 @@
 package org.teamapps.demolessons.p1_intro.l13_tree;
 
-import com.google.common.io.Files;
 import org.teamapps.demolessons.DemoLesson;
 import org.teamapps.icon.material.MaterialIcon;
 import org.teamapps.server.jetty.embedded.TeamAppsJettyEmbeddedServer;
@@ -9,37 +8,35 @@ import org.teamapps.ux.component.absolutelayout.Length;
 import org.teamapps.ux.component.dummy.DummyComponent;
 import org.teamapps.ux.component.flexcontainer.VerticalLayout;
 import org.teamapps.ux.component.panel.Panel;
+import org.teamapps.ux.component.rootpanel.RootPanel;
 import org.teamapps.ux.component.template.BaseTemplate;
 import org.teamapps.ux.component.template.BaseTemplateTreeNode;
 import org.teamapps.ux.component.tree.SimpleTree;
 import org.teamapps.ux.component.tree.SimpleTreeModel;
 import org.teamapps.ux.component.tree.Tree;
 import org.teamapps.ux.session.SessionContext;
-import org.teamapps.webcontroller.SimpleWebController;
+import org.teamapps.webcontroller.WebController;
 
 import java.util.Arrays;
 import java.util.List;
 
 public class TreeDemo implements DemoLesson {
 
-    private Component rootComponent = new DummyComponent();
-    private SessionContext context;
+    private Component rootComponent;
 
-    public TreeDemo(SessionContext context) {
-        this.context = context;
-
+    public TreeDemo(SessionContext sessionContext) {
 
         /* SimpleTree */
         SimpleTree<Void> simpleTree = new SimpleTree<>();
 
-        BaseTemplateTreeNode<Void> locations = new BaseTemplateTreeNode<Void>(MaterialIcon.MAP, "Locations");
+        BaseTemplateTreeNode<Void> locations = new BaseTemplateTreeNode<>(MaterialIcon.MAP, "Locations");
         simpleTree.addNode(locations);
         simpleTree.addNode(new BaseTemplateTreeNode<Void>(MaterialIcon.LOCAL_AIRPORT, "Airport").setParent(locations));
         simpleTree.addNode(new BaseTemplateTreeNode<Void>(MaterialIcon.LOCAL_FLORIST, "Florist").setParent(locations));
         simpleTree.addNode(new BaseTemplateTreeNode<Void>(MaterialIcon.LOCAL_GROCERY_STORE, "Store").setParent(locations));
         simpleTree.setTemplatesByDepth(BaseTemplate.LIST_ITEM_MEDIUM_ICON_SINGLE_LINE, BaseTemplate.LIST_ITEM_SMALL_ICON_SINGLE_LINE);
 
-        simpleTree.onNodeSelected.addListener(node -> context.showNotification(MaterialIcon.MAP, node.getCaption()));
+        simpleTree.onNodeSelected.addListener(node -> sessionContext.showNotification(MaterialIcon.MAP, node.getCaption()));
 
         /* Tree with a Model */
         SimpleTreeModel<Void> simpleTreeModel = new SimpleTreeModel<>();
@@ -65,12 +62,12 @@ public class TreeDemo implements DemoLesson {
     }
 
     private static List<BaseTemplateTreeNode<Void>> createNodes() {
-        BaseTemplateTreeNode<Void> meansOfTransport = new BaseTemplateTreeNode<Void>(MaterialIcon.DIRECTIONS, "Travel by Manpower");
+        BaseTemplateTreeNode<Void> meansOfTransport = new BaseTemplateTreeNode<>(MaterialIcon.DIRECTIONS, "Travel by Manpower");
         BaseTemplateTreeNode<Void> walking = new BaseTemplateTreeNode<Void>(MaterialIcon.DIRECTIONS_WALK, "Walking")
                 .setParent(meansOfTransport);
         BaseTemplateTreeNode<Void> running = new BaseTemplateTreeNode<Void>(MaterialIcon.DIRECTIONS_RUN, "Running").setParent(meansOfTransport);
         BaseTemplateTreeNode<Void> bike = new BaseTemplateTreeNode<Void>(MaterialIcon.DIRECTIONS_BIKE, "Bike").setParent(meansOfTransport);
-        BaseTemplateTreeNode<Void> publicTransportation = new BaseTemplateTreeNode<Void>(MaterialIcon.DIRECTIONS_TRANSIT, "Public Transportation");
+        BaseTemplateTreeNode<Void> publicTransportation = new BaseTemplateTreeNode<>(MaterialIcon.DIRECTIONS_TRANSIT, "Public Transportation");
         BaseTemplateTreeNode<Void> bus = new BaseTemplateTreeNode<Void>(MaterialIcon.DIRECTIONS_BUS, "BUS").setParent(publicTransportation);
         BaseTemplateTreeNode<Void> taxi = new BaseTemplateTreeNode<Void>(MaterialIcon.DIRECTIONS_CAR, "Taxi").setParent(publicTransportation);
         BaseTemplateTreeNode<Void> ship = new BaseTemplateTreeNode<Void>(MaterialIcon.DIRECTIONS_BOAT, "Ship").setParent(publicTransportation);
@@ -97,14 +94,20 @@ public class TreeDemo implements DemoLesson {
     public void handleDemoSelected() { }
 
 
+    // main method to launch the Demo standalone
     public static void main(String[] args) throws Exception {
+        WebController controller = sessionContext -> {
+            RootPanel rootPanel = new RootPanel();
+            sessionContext.addRootPanel(null, rootPanel);
 
-        SimpleWebController controller = new SimpleWebController(context -> {
+            // create new instance of the Demo Class
+            DemoLesson demo = new TreeDemo(sessionContext);
 
-            TreeDemo textFieldDemo = new TreeDemo(context);
-            textFieldDemo.handleDemoSelected();
-            return textFieldDemo.getRootComponent();
-        });
-        new TeamAppsJettyEmbeddedServer(controller, Files.createTempDir()).start();
+            // call the method defined in the DemoLesson Interface
+            demo.handleDemoSelected();
+
+            rootPanel.setContent(demo.getRootComponent());
+        };
+        new TeamAppsJettyEmbeddedServer(controller).start();
     }
 }
